CHROMIUM: drivers/backlight: delay backlight resume a bit
There is a race on resume between the backlight and drm drivers. If
the drm driver wins, the display is initialized and then the backlight
turns on, which looks fine. If the backlight wins, then an unitialized
display is visible and there is a user-visible flash on the screen.

This works around the resume ordering issue by postponing the
backlight resume call for a little while.

BUG=chromium:353667
TEST=by hand: resume daisy, the screen doesn't flash

Change-Id: I0703c00fad08c99899a13a0cc74c6361e24a0ad9
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/190832
Reviewed-by: Olof Johansson <olofj@chromium.org>