drm/exynos: return the number of modes in get_modes()
get_modes is supposed to return a positive value or 0, not an error.
That's what the DRM expects later and checks against 0 to decide to
add the fallback modes. So when we return -ENODEV for example, it
doesn't match since we test 0, and the DRM adds no modes at all.

Also fix the return value from drm_add_edid_modes (we should test
against zero, not against negative values which can't happen).

This fixes the case where the EDID can't be read (either because
things don't work, or because there is no DDC line).

BUG=none
TEST=by hand on snow: I now see some modes with an adapter that doesn't have a DDC line

Change-Id: I86ca14d8f8b25ae88cfe2a30f920328c09a5d087
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/191194
Reviewed-by: David Riley <davidriley@chromium.org>
Reviewed-by: Dominik Behr <dbehr@chromium.org>