CHROMIUM: ARM: Disable use of virtual arch timer registers
We have to use the physical timer instead of the virtual one because
there is a per-cpu offset (CNTVOFF) that is automatically added to the
physical timer value when it is read.  Unfortunately, CNTVOFF has an
unknown reset value and the only way to set it is through hypervisor
mode.  It is a lot simpler in our case to just disable the virtual timer
globally since we already know the ARM platforms on which this kernel
will run.  However, we will need a more elegant solution for
upstream (perhaps some way to disable it in the device tree?).

BUG=chrome-os-partner:13805
TEST=Built and booted the kernel on pit, pi, snow, and spring

Change-Id: I8f55ecbe1443cd42bed1908c2dffa4415afd1a25
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/191608
Reviewed-by: Olof Johansson <olofj@chromium.org>