ARM: exynos542x: Adjust L2 cache timings at resume time
On the exynos5420 and exynos5422 we need to make sure that the Data
RAM Setup and Tag RAM Setup are both set to 1 in order for things to
be stable.  Unfortunately some early firmware shipped with devices
doesn't set this up properly.

We'll tweak the platform-specific resume code to detect exynos542x and
fix.

BUG=chrome-os-partner:27270
TEST=See I1c8d3c48b2a6a617f71a965070aa3d8be6cad887 (with all patches)

Change-Id: Ib4915d1bb05cf4f62328ee28881e31759116b502
Signed-off-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Arjun.K.V <arjun.kv@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/191698