drm/anx7808: Add VGA dongle support
This patch adds VGA dongle support by using SINK_COUNT instead of
DOWN_STREAM_STATUS to determine whether there is a sink connected
to the dongle. In the case of VGA dongles, DOWN_STREAM_STATUS always
reports there is nothing connected, whereas SINK_COUNT shows 1 when
VGA (and HDMI) are present and 0 when they're not.

BUG=chrome-os-partner:26205
TEST=Tested on spring with VGA and HDMI dongles

Change-Id: I70a93c31f709718e0808d6c0ab4fb6600fef6af4
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/191711
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>