drm/anx7808: Detect when bridge fails hotplug and try to recover
On some hotplugs the bridge would fully enable HDMI and then have
the downstream HPD flap.  Upon detecting the loss of downstream HPD
it would be propogated upstream and result in the hotplug failing to
display video on the external display.  This change tracks when HDMI is
fully enabled but avoids disabling upstream HPD and instead relies on
anx7808_detect in determining when the connection is lost.

BUG=chromium:348169
TEST=Tested on spring with HDMI by repeatedly hotplugging HDMI, toggling between adapter being powered and unpowered between hotplugs.

Change-Id: I893bf896522916735435113c346cd85ad115290e
Signed-off-by: David Riley <davidriley@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/192431
Reviewed-by: Sean Paul <seanpaul@chromium.org>