CHROMIUM: config: enable UHID
UHID provides a mechanism for userspace to implement HID I/O Drivers,
performing the raw data I/O to the low level device and pushing HID
Reports to the kernel for parsing.

It's required for HID over GATT (HID on Bluetooth Low Energy) because,
unlike HID on classic Bluetooth, the GATT protocol is implemented in
userspace whereas the L2CAP protocol on which HID is based is
implemented in the kernel.

BUG=chromium:345982
TEST=`emerge-$BOARD chromeos-kernel-next` and check for /dev/uhid

Change-Id: Id7ae4efdd9045fad2eaac3bf84841ff2e3e02102
Signed-off-by: Scott James Remnant <keybuk@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/192415
Reviewed-by: Kees Cook <keescook@chromium.org>