MALI: Refcount mmap vma
Given a mapped GPU managed memory allocation in the address space of a
user process, the user process may perform operations on the mapped
memory that can result into the kernel creating more than one
vm_area_struct (vma) data structure to manage the virtual address range.

The problem is that the kernel driver is only aware of the initial vma
that was created for the allocation and will free the pages allocated
when the kernel notifies the kernel driver that the original vma got
closed (e.g. as result of unmap or mprotect) while other vmas are still
in existence.

BUG=chrome-os-partner:27634
TEST=Sanity check kernel build and operation, verify Mali driver
     operation (sanity check UI).

Change-Id: Ia49af3fce9706d322886dba095960190170fb6c5
Signed-off-by: Ronald Sloot <Ronald.Sloot@arm.com>
Signed-off-by: David Garbett <david.garbett@arm.com>
Reviewed-on: https://chromium-review.googlesource.com/193749
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>