HACK: exynos5420: devfreq: Set wcore frequency to 333 Mhz for all levels
On pit, video decode latency is high when wcore freq is lower.
Since voltage for all levels are the same it is safe to set
wcore frequency to 333 for all the 4 levels.

This can be reverted once a proper fix to lock devfreq lands.

BUG=chrome-os-partner:27529
TEST=Run vdau and measure decode time for 1080p vp8 stream

Change-Id: I99efe0a87f58d0700d938778bbb303e38c3a0820
Signed-off-by: Arun Mankuzhi <arun.m@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/193003
Reviewed-by: Arun Mankuzhi <arun@chromium.org>
Tested-by: Arun Mankuzhi <arun@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Arun Mankuzhi <arun@chromium.org>