CHROMIUM: mwifiex: balance dma map/unmap sizes
Depending on the underlying DMA implementation its
not possible to partially unmap DMA buffers. Moreover
its not possible to understand the intent of passing
0 as the size to dma unmap. The intent of this
driver is unmap the entire skb buffer. The only way
to ensure that the size matches on unmap is to store
both the dma address and the size in the skb ca field.

Introduce a mwifiex_dma_mapping structure which tracks
the dma address and the size. Additionally, provide
a mwifiex_unmap_pci_memory() that utilizes the new
structure. This also provide symmetry within the
internal API.

BUG=chrome-os-partner:24817
BUG=chrome-os-partner:24397
BRANCH=None
TEST=Running with instrumented kernel reveals no more
     0-sized DMA unmaps. And the factory Stress test
     can now pass.

Change-Id: Ic2c076c758503d4242db195433eca146c5c7ee11
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Bing Zhao <bzhao@marvell.com>
Reviewed-on: https://chromium-review.googlesource.com/193350
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Bing Zhao <bzhao@marvell.com>
Tested-by: Bing Zhao <bzhao@marvell.com>