drm/i915: Lock the config_mutex during resume modeset
The warning for the config_mutex was introduced in this backport:

commit 0ef7dfb5ac7bf042ca8e346c8ebeda4cd6827553
Author: Daniel Vetter <daniel.vetter@ffwll.ch>
Date:   Sun Dec 2 01:05:46 2012 +0100

UPSTREAM: drm/i915: use drm_modeset_lock_all

But the locking wasn't done, resulting in a warning. While this a
harmless warning (nothing should contend the lock while we are on the
resume path) we should still do proper locking to make the warning go
away.

BUG=chromium:352453
TEST=compiles, run, suspend/resume on link, no warning observed

Change-Id: I0508020c28c8bb384e059a1966bfe4eaa8d5ce25
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/193479
Reviewed-by: Sean Paul <seanpaul@chromium.org>