UPSTREAM: r8152: check RTL8152_UNPLUG
When the device is unplugged, the driver would try to disable the
device. Add checking the flag of RTL8152_UNPLUG to skip setting
the device when it is unplugged. This could shorten the time of
unloading the driver.

Signed-off-by: Hayes Wang <hayeswang@realtek.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 6871438cc4e5307ccda70fa2a246a546300ac9fa)

BUG=chrome-os-partner:28019
TEST=manual testing by closing/re-open lid with video clip
verify video does not have disconnection when playing

Change-Id: Ic9efe45c7fd500389af89a759c5adaefd0b6ec4a
Signed-off-by: Wonjoon Lee <woojoo.lee@samsung.com>
Signed-off-by: Grant Grundler <grundler@chromum.org>
Reviewed-on: https://chromium-review.googlesource.com/194558
Reviewed-by: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>