CHERRY-PICK: Bluetooth: Fix triggering BR/EDR L2CAP Connect too early
Commit 1c2e004183178 introduced an event handler for the encryption key
refresh complete event with the intent of fixing some LE/SMP cases.
However, this event is shared with BR/EDR and there we actually want to
act only on the auth_complete event (which comes after the key refresh).

If we do not do this we may trigger an L2CAP Connect Request too early
and cause the remote side to return a security block error.

Signed-off-by: Johan Hedberg <johan.hedberg@intel.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Cc: stable@vger.kernel.org
Signed-off-by: Scott James Remnant <keybuk@chromium.org>
BUG=chromium:383984
TEST=emerge, test_that suite:bluetooth

Change-Id: Iafb93b6e5aa385e4c4a0e3ec26b53ebcf80d25f1
Reviewed-on: https://chromium-review.googlesource.com/203754
Reviewed-by: Arman Uguray <armansito@chromium.org>
Commit-Queue: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>