drm/ptn3460: Stop using edid emulation
There have been reports from the field that older devices are suffering
lcd bloom on boot. Although I cannot reproduce it, it's plausible that
the older hardware is failing when reading the emulated edid and we are
failing to get the modes from hardware.

This patch removes the edid emulation and just uses a devicetree supplied
mode.

BUG=chrome-os-partner:28162
TEST=Tested on my snow, screen came up

Change-Id: Iad1acedaa897c3a90a3d143361964aad96914bfa
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/195466
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>