ARM: exynos542x: Re-adjust the L2 cache timings at resume
On the exynos5420 and exynos5422 we need to make sure that the Data
RAM Setup is 1, Tag RAM Setup is set to 1, and Tag and Data latency
is increased by 1, in order for things to be stable.
Unfortunately some early firmware shipped with devices
doesn't set this up properly.
We'll tweak the platform-specific resume code to detect exynos542x and
fix.

BUG=chrome-os-partner:27729
TEST=Run reboot tests on pit and pi and see that there are no hangs.

Change-Id: I85030b32c3d5a79789926e1efab8e16c04dd68c5
Signed-off-by: Arjun.K.V <arjun.kv@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/195813
Reviewed-by: Doug Anderson <dianders@chromium.org>