ARM: dts: Disable the ADC on exynos5250 by default
Upstream the ADC is diabled by default.  Let's do the same so that we
don't need to add an explicit "disabled" for all of our devices.

Note that we used to use the ADC on exynos5250-snow but we stopped
using it since newer revisions of hardware removed their thermistors.
I'll totally remove the ADC from the snow device tree in the next
change.

BUG=chrome-os-partner:27407
TEST=Boot and see no adc errors in logs; see no other problems

Change-Id: Iffd6f9c695fa3f2372ca7ca716057e6d47932c81
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/196253
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>