PM / Sleep: arm: mmu: Fix fixed pte suspend volatile mapping
In (31c22cd UPSTREAM: arm: use generic fixmap.h) we changed to using
the generic fixmap implementation.  This has the notable side effect
that fixmap id 0 is now the largest address, not the smallest.

Adjust the old (PM / Sleep: arm: mmu: Mark fixed ptes as volatile
during suspend) patch to handle the new fixmap style.

BUG=chromium:365011
TEST=Print volatile chunk base addr / size before and after this
series and see that it is identical.

Change-Id: I0db0ddfb00d9deee7a2cf9f715ec164e28a0427a
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/196962
Reviewed-by: Kees Cook <keescook@chromium.org>