ARM: exynos542x: Mask the watchdog at early kernel boot
The bootloader on some exynos542x systems (like peach-pit and
peach-pi) leaves the watchdog on to detect problems that might happen
really early in the transition to Linux.

Previously, the way that this used to work was:
1. Bootloader left the watchdog on.
2. Early kernel boot code would pat the watchdog but leave it enabled.
3. Roughly 3.5 seconds into boot the watchdog driver would start up
   and disable the watchdog.
4. After userspace started up, the watchdog would start again.

The problem comes in when we disable the watchdog driver.  This is a
sensible thing to do and is what happens with the kgdb USE flag after
<https://chromium-review.googlesource.com/196642>.  That made the
flow:
1. Bootloader left the watchdog on.
2. Early kernel boot code would pat the watchdog but leave it enabled.
3. Roughly 5 seconds into boot the watchdog clock would get gated,
   effectively disabling the watchdog (though in an odd way).
4. When the device tried to go into suspend state, the watchdog would
   fire.

Given that the old state was only really "protecting" the system for
about 3.5, let's just change the early kernel to mask the watchdog.
We still get protection for the U-Boot to Linux transition but we lose
protection of the 3.5 seconds of early boot.

BUG=chromium:365011
TEST=With USE=kgdb, suspend_stress_test works
TEST=Without USE=kgdb, "stop daisydog; cat /dev/watchdog0" reboots
TEST=Without USE=kgdb, suspend_stress_test works

Change-Id: I3db1f3272d6b9f99556695cf1b1de4dfef72ff1c
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/197091
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>