UPSTREAM: brcmfmac: Fix reconnect failure after beacon timeout
The DISASSOC command needs to be sent to firmware when a connection
loss is detected by firmware (e.g., beacon timeout). Otherwise the
next connect request fails due to a lingering LINK(down) event from
firmware. This patch resolves the issue by eliminating the incomplete
duplicated if statement and using brcmf_link_down() to take care of
the event.

BUG=chrome-os-partner:28366
TEST=autotest (network_WiFi_MissingBeacons)

Change-Id: I459f9cf1cb051ec54ad5dd8f9830f84a6f167e75
Signed-off-by: Daniel Kim <dekim@broadcom.com>
Reviewed-on: http://hnd-swgit.sj.broadcom.com:8080/322
Reviewed-by: brcm80211 ci
Reviewed-by: Arend Van Spriel <arend@broadcom.com>
Reviewed-by: Pieter-Paul Giesberts <pieterpg@broadcom.com>
Reviewed-by: Hante Meuleman <meuleman@broadcom.com>
Reviewed-on: https://chromium-review.googlesource.com/197065
Reviewed-by: Paul Stewart <pstew@chromium.org>