UPSTREAM: usb: xhci: Correct last context entry calculation for Configure Endpoint
The current XHCI driver recalculates the Context Entries field in the
Slot Context on every add_endpoint() and drop_endpoint() call. In the
case of drop_endpoint(), it seems to assume that the add_flags will
always contain every endpoint for the new configuration, which is not
necessarily correct if you don't make assumptions about how the USB core
uses the add_endpoint/drop_endpoint interface (add_flags only contains
endpoints that are new additions in the new configuration).

Furthermore, EP0_FLAG is not consistently set in add_flags throughout
the lifetime of a device. This means that when all endpoints are
dropped, the Context Entries field can be set to 0 (which is invalid and
may cause a Parameter Error) or -1 (which is interpreted as 31 and
causes the driver to keep using the old, incorrect value).

The only surefire way to set this field right is to also take all
existing endpoints into account, and to force the value to 1 (meaning
only EP0 is active) if no other endpoint is found. This patch implements
that as a single step in the final check_bandwidth() call and removes
the intermediary calculations from add_endpoint() and drop_endpoint().

Signed-off-by: Julius Werner <jwerner@chromium.org>
(cherry picked from commit 61cee971a20c392f89e153c1ec5bf8e714cd5245)
Cherry-picked from mnyman/xhci.git/for-usb-linus with no conflicts.

BUG=None
TEST=Run with XHCI debugging, connect/unplug/S2R some devices that like
to change configurations and alternate settings (e.g. cameras). Make
sure the new Slot Context always contains the right Context Entries
value after each check_bandwidth().

Change-Id: I9cdef7367a8b39365e9b008872214881f0d0dd6c
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/199813
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>