CHROMIUM: gsc: Do not adjust buffer size for buffers not allocated by us.
To allow using/sharing buffers allocated elsewhere, we should not be
adjusting their size (if one is provided) when they are queued.

Additionally, if userspace requests larger buffers than what we would
require to fit the selected format, honor it and do not adjust to our
requirements, unless it would be too small to fit the data we would
produce. In that case adjust it on S/TRY_FMT to indicate the size would
be too small.

This requires separating payload size from buffer size associated with
the frame.

Signed-off-by: Pawel Osciak <posciak@chromium.org>

BUG=chromium:345179
TEST=veatest

Change-Id: Ibc463281b7c22ad28762ca0cee2cbd832bb896be
Reviewed-on: https://chromium-review.googlesource.com/198662
Reviewed-by: Pawel Osciak <posciak@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>