CHROMIUM: s5p-mfc: fix a race in interrupt flags handling
Interrupt result flags have to be cleared before a hardware job is run.
Otherwise, if they are cleared asynchronously, we may end up clearing them
after the interrupt for which we wanted to wait has already arrived, thus
overwriting the job results that we intended to wait for.

To prevent this, clear the flags only under hw_lock and before running
a hardware job.

Signed-off-by: Pawel Osciak <posciak@chromium.org>

BUG=chromium:370758
TEST=vdatest, veatest

Change-Id: Ia00a2525f885795a934feb6c282784865e0212c8
Reviewed-on: https://chromium-review.googlesource.com/198803
Reviewed-by: Kiran AVND <kiran@chromium.org>
Reviewed-by: Wu-Cheng Li <wuchengli@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>