CHROMIUM: s5p-mfc: Handle first buffer not containing a stream header.
Currently, when initializing the decoder, it is required that the first
stream buffer queued contains a stream header, otherwise decoding will
fail. The hardware however does not have this limitation, it will return
an error indicating that there was no header in the buffer, but will allow
continuing until a subsequent buffer with a valid header is queued.

Add handling of this situation to the driver and allow processing more
input stream buffers until a valid header is found.

Signed-off-by: Pawel Osciak <posciak@chromium.org>

BUG=chromium:370758
TEST=veatest, vdatest

Change-Id: Ifb1ca2e37e4b688949023e65f177acc5aeeee393
Reviewed-on: https://chromium-review.googlesource.com/198804
Reviewed-by: Wu-Cheng Li <wuchengli@chromium.org>
Commit-Queue: Pawel Osciak <posciak@chromium.org>
Tested-by: Pawel Osciak <posciak@chromium.org>