CHROMIUM: iwl7000: mac80211: mark local variable __maybe_unused
The 'local' variable in __ieee80211_vif_copy_chanctx_to_vlans()
is only used/needed when lockdep is compiled in, mark it as such
to avoid compile warnings in the other case.

While at it, fix some indentation where it's used.

Remove the WARN_ON() that was previously used to suppress the
compiler warning (but actually generates code.)

Change-Id: I6eb29751b1094b8bf719fecd8d5207a57501acdf
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/200399
Reviewed-by: Paul Stewart <pstew@chromium.org>