CHROMIUM: devfreq: exynos5422: dynamically handle HDMI
We don't want to pay the voltage penalty required for HDMI at all
times, so we'll detect when HDMI is enabled and only bump up the INT
voltage then.  This means that _all_ INT clocks will be bumped up to
LV_2 when HDMI is plugged in, but hopefully that's a use case where a
little extra power draw is expected.

BUG=chrome-os-partner:26245
TEST=performance, HDMI, video decode, stability, suspend_stress

Change-Id: I916216f6bbbcf71e3154460db282ecac624caded
Signed-off-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Arjun.K.V <arjun.kv@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/201043
Reviewed-by: Akshu Agrawal <akshu@chromium.org>