mmc: dw_mmc: Fix warning on dw_mci_request_end()
This fixes a warning that showed up after (62ef5d2 mmc: dw_mmc: Make
sure we don't get stuck when we get an error) and after we enabled
tuning.  The warning seems very common on some particular machines and
doesn't show up at all on other machines.

What appears to be happening is that we get a tuning error and we
detect it properly.  Then we call send_stop_abort() which sends a stop
command and set state to STATE_DATA_ERROR.  We'll move from
STATE_DATA_ERROR to STATE_DATA_BUSY, where we'll find out that the
stop command actually got a 2nd error without getting a CMD_COMPLETE.
We'll run the new code to call dw_mci_request_end() but cmd will be
non-NULL, so it will warn.

It appears to be OK to just NULL out the cmd to get rid of the error.

We'll also get rid of an extra TODO in the code since it seems to be
working fine as-is.

BUG=chrome-os-partner:28672
TEST=suspend_stress_test -c20 on affected machine, no errors

Change-Id: I50045431d5fb10bdc63ecb5b6e734e49e0132284
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/201805