drm/exynos: Add 1366x768 support to the mixer
Now that 1360x768 works, there is no reason not to do 1366x768 as
well since it uses the same clock. This mode requires a hack to
turn it into 1360 though, so we do that here. Thankfully all the
infrastructure for exynos5250 adjustments can be reused so this isn't
so hacky in the end.

BUG=chrome-os-partner:26896
TEST=on pi and daisy, with Dominik's monitor, try 1366x768 and it works

Change-Id: Icb113b06067cb674d7f2716328da5a8044be5959
Reviewed-on: https://chromium-review.googlesource.com/202152
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Dominik Behr <dbehr@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>