ARM: dts: Add the regulator locker node older revs of pit
In (ec3d02f arm: dts: exynos542x: add int locker regulator node) we
added the regulator locker to rev6+ of pit, but we didn't add it to
rev 3, 4, and 5.  Move the change to the dtsi so everyone gets it.

NOTE: I purposely didn't apply this to the "pi" rev 8/9, which I plan
to remove shortly.

BUG=chromium:379700
TEST=Boot on rev9 and see locker get probed in dmesg.

Change-Id: Ifa6e1996672934b96beb9f9bd8c199a598846fa8
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/202346
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>