CHROMIUM: Add support for Logitech TK820
The TK820 is a keyboard that sports a big square touchpad.
This device supports the x6100 hidpp feature for raw data reporting.

Besides adding the Wireless PID to the list of this driver supported
devices, some extra tweaks were needed to convey the keyboard part
of the device.

Parsing of the keyboard part of the hid descriptor is still performed
by hid_core, as well as the forwarding of the keyboard events to the
appropriate input device.

(cherry picked from e6768ccb04c1649355317808396a5534d4930522 in 3.14)

BUG=none
TEST=Tested on 3.14 lumpy hw with a TK820/K830 at Logitech.

Change-Id: I884cce7d629c4be73567de112996a1edda488940
Reviewed-on: https://chromium-review.googlesource.com/212021
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Nestor Lopez Casado <nlopezcasad@logitech.com>
Commit-Queue: Nestor Lopez Casado <nlopezcasad@logitech.com>
Tested-by: Andrew de los Reyes <adlr@chromium.org>