CHERRY-PICK: Bluetooth: Reuse hci_stop_discovery function when cleaning up HCI state
When cleaning up the HCI state as part of the power-off procedure we can
reuse the hci_stop_discovery() function instead of explicitly sending
HCI command related to discovery. The added benefit of this is that it
takes care of canceling name resolving and inquiry which were not
previously covered by the code.

BUG=chrome-os-partner:26547
TEST=Enable and disable the BT device multiple times

Signed-off-by: Johan Hedberg <johan.hedberg@intel.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit 986603b8949d4347717105556f774f956d8fc605)
Signed-off-by: Tedd Ho-Jeong An <tedd.an@intel.com>

Change-Id: Ib2e1399280b9db86997df7612cdc9e114e06c76e
Reviewed-on: https://chromium-review.googlesource.com/203532
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Tested-by: Tedd An <tedd.an@intel.com>