drm/exynos: Call hdcp_disable instead of hdcp_stop
This patch changes the hdmi conf_apply function to call hdcp_disable
instead of hdcp_stop. The reason we need to do this is because the
mode_config.mutex is locked in the commit/dpms_on path and hdcp_stop
synchronously waits for hdcp_work to finish, which also tries to acquire
mode_config.mutex. The result is a deadlock if hdcp is both enabled and
disabled in the modeset path.

BUG=chromium:389282
TEST=Tested on spring with hdcp enable in modeset path

Change-Id: Ife627408fbcdada6bedabc99d75c90dc48583472
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/206023
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>