drm/anx7808: Properly disable/re-enable hdcp on unplug
This patch disables and re-enables HDCP when the HDMI cable is
unplugged and re-plugged. To achieve this with the anx7808's warts, we
disable the upstream encoder's hdcp on disable, and then re-enable it
after a delay on enable. The reason we need a delay on re-enable is b/c
it seems the bridge doesn't like it when HDCP is enabled too close to power
on.

BUG=chromium:389282
TEST=Tested on spring, enable HDCP then unplug HDMI and replug

Change-Id: I2ea611fab2b504f25ff2f542f8715a6f6da374ba
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/206025
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>