CHERRY-PICK: CHROMIUM: Input: atmel_mxt_ts: fix out-of-bounds array accesses
This is cherry-picked from 0fa7c8d4bf70722be1bcd4adb0f54ef832e84449

The current_id map is fixed to a size of MXT_MAX_FINGER (10), but
the touch controller may generate reports for a larger number of
IDs.  On Nyan, for example, the controller claims it can report
up to 16 touch IDs.  During initialization of the touchscreen,
events were generated for all 16 touch IDs causing out-of-bounds
writes to the current_id map.  Instead of fixing the size to 10,
dynamically allocate the current_id map based on the number of
touch IDs the controller claims to support.
BUG=none
TEST=Build and boot on Nyan.  evtest output looks reasonable.
Suspend/resume is no longer broken.
Change-Id: Iabf5f91839ed1523374209beac7daa209bbf060d
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/179758
Reviewed-by: Yufeng Shen <miletus@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/207264
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Yufeng Shen <miletus@chromium.org>
Tested-by: Yufeng Shen <miletus@chromium.org>