drm/exynos: Always setup the primary plane
In the new world of atomic modeset, the primary plane isn't enabled by
default when we set a mode. This breaks a number of things, in
particular framebuffer (which in turn breaks the splash screen and the
recovery install screen).

To fix this, we always enable the primary plane on modeset by calling
update_plane() from commit(). This is limiting the API a bit since
this always forces the primary plane to be enabled, but ensures
compatibility with old user space and framebuffer.

BUG=chromium:390147
TEST=boot on daisy, see a framebuffer

Change-Id: I3e4d522ac5d16b2d38d8031fdb42ae65558f965f
Reviewed-on: https://chromium-review.googlesource.com/207708
Tested-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stéphane Marchesin <marcheu@chromium.org>