drm/exynos: Invalidate crtc mode when powered off
This patch invalidates the crtc's current mode when we enter set_config with
the hardware powered off. This will force the crtc helper to go through a full
modeset instead of short-circuiting with mode_set_base. By doing this, we can
now remove the dpms on call from mode_set_base since that function should really
only be called with the hardware is active (since this has the effect of turning
on the crtc without the encoders/bridges powered).

BUG=None
TEST=Tested on peach-pit dpms/suspend work as expected

Change-Id: I5d2946a837333953b28b3b5e885842de8a0b7446
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/208340
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>