drm/exynos: Call hdcp_disable instead of hdcp_stop in poweroff
This patch changes the poweroff routine to call hdcp_disable instead of
hdcp_stop. The reason we need to do this is that both the disable
path and hdcp_work try to acquire mode_config.mutex. By calling hdcp_stop,
we synchonously wait for hdcp_work to complete while holding the
mode_config.mutex while the worker waits on it. Hence, deadlock.

BUG=chromium:394400
TEST=Tested on peach, can logoff with HDCP enabled

Change-Id: I8f1dff08226e40bbdd7845fc4bcdda2cf81b0867
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/209301
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>