BACKPORT: clocksource: Move cycle_last validation to core code
The only user of the cycle_last validation is the x86 TSC. In order to
provide NMI safe accessor functions for clock monotonic and
monotonic_raw we need to do that in the core.

We can't do the TSC specific

    if (now < cycle_last)
       	    now = cycle_last;

for the other wrapping around clocksources, but TSC has
CLOCKSOURCE_MASK(64) which actually does not mask out anything so if
now is less than cycle_last the subtraction will give a negative
result. So we can check for that in clocksource_delta() and return 0
for that case.

Implement and enable it for x86

BUG=chromium:324025
TEST=build/boot 3.8 on link

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: John Stultz <john.stultz@linaro.org>
(cherry picked from commit 09ec54429c6d10f87d1f084de53ae2c1c3a81108)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
[SR: trivial conflict in Kconfig]
[SR: 3.10: fix conflict in timekeeping_internal.h]
[SR: 3.8: fix context conflict in tsc.c from 82f9c080]
Conflicts:
	arch/x86/Kconfig
	arch/x86/kernel/tsc.c
	kernel/time/timekeeping_internal.h

Change-Id: Ib980b5ef4733e9f73f8c5222221ad439f8b306f6
Reviewed-on: https://chromium-review.googlesource.com/300773
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>