BACKPORT: timekeeping: Create struct tk_read_base and use it in struct timekeeper
The members of the new struct are the required ones for the new NMI
safe accessor to clcok monotonic. In order to reuse the existing
timekeeping code and to make the update of the fast NMI safe
timekeepers a simple memcpy use the struct for the timekeeper as well
and convert all users.

BUG=chromium:324025
TEST=build/boot 3.8 on link

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Ingo Molnar <mingo@kernel.org>
Cc: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Signed-off-by: John Stultz <john.stultz@linaro.org>
(cherry picked from commit d28ede83791defee9a81e558540699dc46dbbe13)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
[SR: several non-trivial differences from refactoring timekeeping.c
     and vsyscall code in x86 vsyscall_64.c -> vsyscall_gtod.c]
[SR: 3.8: tile doesn't have vsyscall in 3.8 and fix up timekeeper_resume()]
Conflicts:
	arch/s390/kernel/time.c
	arch/tile/kernel/time.c
	arch/x86/kernel/vsyscall_gtod.c
	arch/x86/kvm/x86.c
	include/linux/timekeeper_internal.h
	kernel/time/timekeeping.c

Change-Id: Ie11521ead23722087f03e4ec3538cbdf34dc943a
Reviewed-on: https://chromium-review.googlesource.com/300780
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>