CHROMIUM: mac8011: provide debugfs file to wake queues
On devices that use the ath9k driver, we've observed some cases
where the mac80211 TX queues get stuck in a stopped state. This
case is surprising, since the TX queues are stopped due to
power-save mode, even though we are not associated with an AP.

We don't know why that's happening. For now, we want to work
around the issue, rather than root-cause it. To that end, this
CL adds a debugfs file that userspace can use to forcibly wake
queues.

Signed-off-by: quiche@chromium.org

BUG=chromium:392209
TEST=compile

Change-Id: I90f24dafe58ded211f6af3356693b8e4a5470e5d
Reviewed-on: https://chromium-review.googlesource.com/211712
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>