CHROMIUM: usb: cdc-wdm: Clear read pipeline in case of error.
commit 73e06865ead1 ("USB: cdc-wdm: support back-to-back
USB_CDC_NOTIFY_RESPONSE_AVAILABLE notifications") implemented queued response
handling. This queue is processed every time the WDM_READ flag is cleared.

In case of a read error, userspace may not actually read the data, since the
driver returns an error through wdm_poll. After this, the underlying device may
attempt to send us more data, but the queue is not processed. While userspace is
also blocked, because the read error is never cleared.

After this patch, we proactively process the queue on a read error. If there was
an outstanding response to handle, that will clear the error (or go through the
same logic again, if another read error occurs). If there was no outstanding
response, this will bring the queue size back to 0, unblocking a future response
from the underlying device.

BUG=chrome-os-partner:30325
TEST=Run network_3GModemControl till we see -EPIPE from cdc-wdm driver. Ensure
that the system recovers, and the test passes.

Change-Id: I6ba444fe03c903b90186e9550c5ff145a93ca6fa
Signed-off-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/219924
Reviewed-by: Ben Chan <benchan@chromium.org>